🔥 completly removed bot from this repo
