+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "nerddiary"
 version = "0.1.0"
-description = "A collection of tools to capture and analyze migraines"
+description = "A collection of tools to capture a personal log / diary and analyze these records"
 authors = ["mishamsk <mishamsk@gmail.com>"]
 license = "Apache-2.0"
 readme = "README.md"
@@ -31,9 +31,6 @@ jsonrpcserver = { version =  "^5.0.6", optional = true }
 click = { version = "^8.0.3", optional = true }
 SQLAlchemy = { version = "^1.4.29", optional = true }
 cryptography = { version = "^36.0.1", optional = true }
-cryptg = { version = "^0.2.post4", optional = true }
-Pillow = { version = "^9.0.0", optional = true }
-Telethon = { version = "^1.24.0", optional = true }
 APScheduler = { version = "^3.8.1", optional = true }
 uvicorn = { version = "^0.17.4", optional = true }
 fastapi = { version = "^0.73.0", optional = true }
@@ -41,8 +38,6 @@ fastapi = { version = "^0.73.0", optional = true }
 [tool.poetry.extras]
 full = ["jsonrpcclient", "jsonrpcserver", "click", "SQLAlchemy","cryptography","cryptg","Pillow","Telethon","APScheduler","uvicorn","fastapi"]
 client = ["jsonrpcclient"]
-# Must include all client stuff
-tgbot = ["jsonrpcclient", "click", "cryptg", "Pillow", "Telethon"]
 server = ["jsonrpcserver", "click", "SQLAlchemy", "cryptography", "APScheduler", "uvicorn", "fastapi"]
 
 [tool.poetry.dev-dependencies]
@@ -80,7 +75,6 @@ exclude_lines = [
 
 [tool.black]
 line-length = 120
-target-version = ['py310']
 
 [tool.isort]
 profile = "black"
